feat: Wish Model, Interface 내 좋아요 추가
