--- src/interfaces/IBoardInterestList.ts
@@ -1,6 +1,0 @@
-import mongoose from "mongoose";
-
-export interface IBoardInterestList {
-    member: mongoose.Types.ObjectId;
-    choice: string;
-}
