--- src/interfaces/IBoardPostList.ts
@@ -1,7 +1,0 @@
-import mongoose from "mongoose";
-
-export interface IBoardPostList {
-    poster: mongoose.Types.ObjectId;
-    tag: string;
-    content: string;
-}
