--- src/interfaces/IWishList.ts
@@ -1,7 +1,0 @@
-import mongoose from "mongoose";
-
-export interface IWishList {
-    poster: mongoose.Types.ObjectId;
-    tag: string;
-    content: string;
-}
